refactor: use 'readonly' for injected services and improve import paths for consistency
