+++ src/app/domains/info/pages/home/home.component.ts
@@ -2,11 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { Title } from '@angular/platform-browser';
 import { CommonModule } from '@angular/common';
 import { RouterLink } from '@angular/router';
-import { TranslateService } from '@ngx-translate/core';
-
+import { TranslateService, TranslateModule } from '@ngx-translate/core';
 import { CarouselModule } from 'ngx-bootstrap/carousel';
-import { TranslateModule } from '@ngx-translate/core';
-
 import { NgxBootstrapModule } from '../../../shared/modules/ngx-bootstrap.module';
 import { Project } from '../../../shared/interfaces/project';
 import { ProjectsService } from '../../../shared/services/projects.service';
@@ -22,17 +19,17 @@ export default class HomeComponent implements OnInit {
 
   project = {} as Project;
   years: number;
-  private startDate: Date = new Date(2012, 11, 16);
+  private readonly startDate: Date = new Date(2012, 11, 16);
   //En JavaScript, los meses se indexan desde 0 (enero es 0, febrero es 1, etc.).
-  //Fecha de incio de labores (16 de diciembre del 2012).
+  //Fecha de inicio de labores (16 de diciembre del 2012).
 
   aboutMeTexts: AboutMeTexts = { texts: [] };
 
   constructor(
-    private titleService: Title,
-    private projectService: ProjectsService,
+    private readonly titleService: Title,
+    private readonly projectService: ProjectsService,
     public ngxBootstrapModule: NgxBootstrapModule,
-    private translate: TranslateService
+    private readonly translate: TranslateService
   ) {
     this.titleService.setTitle('Home');
     this.years = this.calculateYears(new Date(this.startDate));
