+++ src/app/domains/info/pages/not-found/not-found.component.ts
@@ -11,7 +11,7 @@ import { RouterLink } from '@angular/router';
 })
 export default class NotFoundComponent {
 
-  constructor (private titleService: Title) {
+  constructor (private readonly titleService: Title) {
     this.titleService.setTitle('404')
   }
