+++ src/app/domains/shared/components/footer/footer.component.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 
 import { TranslateModule } from '@ngx-translate/core';
 
-import { MenuItemContact } from '../../../shared/interfaces/menu-item';
+import { MenuItemContact } from '../../interfaces/menu-item';
 
 @Component({
   selector: 'app-footer',
