+++ src/app/domains/shared/components/header/header.component.ts
@@ -3,8 +3,8 @@ import { CommonModule } from '@angular/common';
 
 import { TranslateModule } from '@ngx-translate/core';
 
-import { TranslationService } from '../../../shared/services/translation.service';
-import { MenuItemHeader, ItemTitle } from '../../../shared/interfaces/menu-item';
+import { TranslationService } from '../../services/translation.service';
+import { MenuItemHeader, ItemTitle } from '../../interfaces/menu-item';
 
 @Component({
   selector: 'app-header',
@@ -25,7 +25,7 @@ export class HeaderComponent {
     { flag: 'Spain Flag', img: '../../../../../assets/icons/spain.png', event: () => this.changeLanguage('es') },
   ]);
 
-  constructor( private translationService: TranslationService ) {}
+  constructor( private readonly translationService: TranslationService ) {}
 
   changeLanguage(lang: string): void {
     this.translationService.changeLanguage(lang);
