+++ src/app/domains/shared/components/navbar/navbar.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit, signal } from '@angular/core';
 import { TranslateService, TranslateModule } from '@ngx-translate/core';
 import { RouterModule } from '@angular/router';
 import { CommonModule } from '@angular/common';
-import { MenuItemNav } from '../../../shared/interfaces/menu-item';
+import { MenuItemNav } from '../../interfaces/menu-item';
 
 @Component({
   selector: 'app-navbar',
@@ -15,7 +15,7 @@ import { MenuItemNav } from '../../../shared/interfaces/menu-item';
 export class NavbarComponent implements OnInit {
   public menuItemsHome = signal<MenuItemNav[]>([]);
 
-  constructor(private translate: TranslateService) {}
+  constructor(private readonly translate: TranslateService) {}
 
   ngOnInit(): void {
     this.translate.get('NAVIGATION').subscribe((navigation: any) => {
