+++ src/app/domains/shared/services/projects.service.ts
@@ -141,7 +141,7 @@ export class ProjectsService {
       let foundAll = true;
 
       filterTags.forEach(function (filterTag) {
-        if (project.tags.includes(filterTag) == false) {
+        if (!project.tags.includes(filterTag)) {
           foundAll = false;
         }
       });
